Move video decoding to it's own thread.
The stream is still read from one thread, but the video decoding has
been moved to it's own thread so it doesn't block the audio decoder
(which is still on the main thread but may soon be moved to it's own
thread). This seems to have fixed the ALSA underuns.