Added another layer of buffering for video
It works like this for both audio and video

[stream] --> [buffer] --> [decoder] --> [buffer] --> [renderer]

We need to make sure now that the video starts and stays synchronized.
Even though the video seems to be playing at the right rate (synced with
the MONOTONIC system clock) it gets out of sync after a while. The audio
is synced by alsa so perhaps we need to use the same clock source
(how?). Or maybe something's wrong with my math.