Sync video to video clock
The video is now synchronized to the audio stream. It worked good on all
my tests but the code used to generate the alsa clock makes assumptions
about the state of the ALSA buffer when the stream is paused (or stalls)
that may not always hold.

This commit also breaks playback of a video without audio.