Video output uses same code with or without video
The video output thread now uses *mostly* the same code to synchronize
the video to a master source. If there is audio then the master source
is the audio clock, if there's no audio it uses the system clock. Two
functions are provided to get the system time and the audio time and
it is set as a function pointer during stream setup.

The code for pausing is still handled within the video output thread but
the idea is to move it to the control functions (mb_player_pause() and
mb_player_play()) and control the video output thread by solely by
modulating the clock source.

This will also allow us to implement fast forward by the following
procedure:

1. Pause the audio thread (mb_player_pauseaudio())
2. Reset the system clock to the value of the audio clock.
3. Multiply the system clock by a speed facor (on
   mb_player_getsystemtime()).
4. The output thread should instruct the decoder to skip frames as
   needed to keep up but we can improve performance by making the stream
   decoder (or reader) thread to skip frames when a speed factor is set.

Rewinding will be trickier.