Real fix for player deadlock
After much hair pulling and trying all kinds of things to fix the video
deadlock I finally figured what was going on. My custom sleep() function
was broken!! Anytime the debugger stopped on it I assume that we where
sleeping, not stuck there. So I assume the audio clock was lagging
behind because I'm looking at my copy of it in local storage which would
get normally get updated on every frame time.