Added the boilerplate for discovery service
On the device end we added a new thread that broadcasts a string
containing the hostname, ip address, and supported features of the
device. On the remote we added a background service that listens for
these broadcasts and maintains a list of available devices.

The device end only broadcasts a bogus IP address at the moment and the
full list of features (since there's no facility to enable/disable
features at the moment).