Make input tcp server multi-threaded
The tcp input server was using a single thread which was fine for now
since we only need one remote. The problem is that if the client device
gets disconnected from the wifi network the connection may not close
immediately and no further connections can be made. This commit
alleviates this problem by running each incoming connection on it's own
thread. It is just a temporary fix until I can research how to properly
detect the broken connection but it also adds support for multiple TCP
input devices.