mediabox: Change UI style/Add cairo support to windows
Changed the fonts and colors and added cairo support to the video
driver. Cairo support is implemented through the function
mbv_window_cairo_begin() and mbv_window_cairo_end() with create and
destroy a cairo_t object for drawing to the window.

Also the about dialog switched from using the video driver to using
pango for rendering text. The plan is to switch everything to pango,
then remove the text functions from the DirectFB driver and implement
them on the generic driver using pango (or at least implement the basic
functionality to get system fonts, etc). This will pave the road for
switching from DirectFB to libdrm.