mediabox: Move mbv_window structure to generic video driver
The old opaque mbv_window structure was renamed to mbv_dfb_video and a
new mbv_window that right now is merely a wrapper around mbv_dfb_video
was defined in the generic video driver (video.c). This is the first
step towards moving all the rendering to the generic driver (except for
accelerated features) and make the device specific driver merely a basic
compositor.