mediabox: Add message based callbacks to timers
Currently the only way to get notifications when a timer expires is
through the callback function. This function is called from the timers
thread so it may require synchronization. This commit adds the ability
to request notifications via the message queue so all processing can be
done in a single thread.

It adds the MBI_TIMER_MESSAGE timer flag and message_fd argument
to mbt_register(). When this flag is set message_fd is the file descriptor
of the message queue where the MBI_EVENT_TIMER messages will be sent
to.