mediabox: Use ALSA non-blocking writes
Since we need to acquire a lock before writing and before pausing the
audio stream we don't want snd_pcm_writei() to block while holding the
lock because then a call to mb_audio_stream_pause() can take several
seconds to complete (depending on the ALSA buffer size and the size of
the audio packet).

This commits solves the problem by setting the pcm handle to
non-blocking mode and sleeping once the ring buffer is full until it gets
half empty.