mediabox: Fix race condition on player
We we're locking the video and audio packet locks only when the buffer
was full but unlocking it every time which is undefined behaviour.

Since the buffer currently only holds 1 packet the lock was being
acquired most of the time, but when it didn't it opens the possibility
of signaling the decoder threads after they acquired the lock but before
going to sleep. In that case the thread would deadlock.

This commit fixes the problem by always acquiring the lock before
checking if the buffer is full.