mediabox: Fix ALSA underrun time calculation
When calculating the ALSA underrun time we called
mb_audio_stream_gettime() to get the current audio time and then get the
amount of audio in the buffer from our own pcm_status_t. So the current
time and the data remaining are calculated at different points in time
leading to a small error that can add up over time and cause the player
to deadlock.

This commit fixes that problem by adding an internal version of
mb_audio_stream_gettime() that also takes in an snd_pcm_status_t pointer
so the current time and the data remaining are taken at the same time.