mediabox: Limit size of ALSA writes
Currently we assume that the size of each decoded audio packet is
smaller than half the ALSA buffer size. This wasn't a problem with
blocking writes because snd_pcm_writei() blocks until the whole packet
has been written in fragments.

With non-blocking writes (commit  0cb6264e2) when a write fails because
the buffer is full we're sleeping until the buffer is half empty, so if
the packet is greater than 1/2 the buffer size we'll end up sleeping
twice and by the time we wake up the second time the buffer will have
underrun already.

This commit fixes the problem by breaking the packets into fragments of
up to half the ALSA period size and when snd_pcm_writei() fails because
the buffer is full we only sleep for the amount of time that will take
to free enough space to write the frame. This Should ensure that the
ALSA buffer never gets much more than half empty.