mediabox: Use a dedicated mutex/cond pair for starting decoder
We were using the same mutex and condition variable used by the decoder
threads to wake the stream parser to signal the start of the stream
parser thread, so if the decoder is fast enough it can signal before the
stream parser.

This commit fixes the problem by using a dedicated mutex and condition
variable to signal that the decoder is ready.