mediabox: Change audio packet size 1/4 buffer size
We were writing packets up to 1/2 the ALSA buffer size. This has a
hight probability of underrun if for example we try to write a packet
when the buffer if 1/2 full plus 1 frame since we go to sleep for 1/2 a
period when the write fails.

Changing the max packet size to 1/4 of the ALSA buffer should really
ensure that the buffer doesn't get much more than 1/2 empty.