mediabox: Implement basic window management on generic driver
Instead of using DirectFB's window manager implement a basic window
manager on the generic driver and use DirectFB only for surface
management. Once this is stable the DirectFB driver will only use basic
features that are also available on libdrm so it will be much easier to
write a libdrm driver and even a cairo based software driver for use
where DirectFB doesn't offer any accelleration.

This is a work in progress. At the time of this commit I have only
verified that the code compiles so it will likely not run.