mediabox: Complete rewrite of video drivers
Now the DirectFB backend is used only for surface
management. All compositing functions are now implemented
on the generic video driver. This is much more efficient
than using DirectFB's window manager. So much so that
we can now use it for rendering video instead of the
fbdev hack that we where using.

NOTE: There's a lot of changes on this commit so it's not
possible to document them all. Aside from the above
mentioned changes a lot of work has been done towards
launching mediatomb, avmount, and deluge without needing
them to be configured priorly but that is a WIP that is
not working. Also the default audio device was switched
to sysdefault since I installed pulseaudio to find out
that our driver ALSA driver doesn't work with the
pulseauio ALSA plugin.

TODO: We need to add support for transparent windows so
we can use the windowing system to display overlay text
and subtitles instead of rendering them on top of each
frame. Also the widgets need to be optimzed to only do
painting when their contents are actually updated. Right
now when a menu is displayed on top of a video the menu
is redrawn every time even though a good copy already
exists on the offscreen surface.