video-drm: Premap surface buffers
Instead of mapping the surface buffers on each lock() call we now map
them when the surface is create. Also fix double-update bug on the
generic driver. When using the directfb driver this only caused wasted
cycles, but with the libdrm driver it causes the wrong buffer to be
flipped (since it's flipped twice so fast).