audio: Fix audio ring buffer corruption issues.
We where using snd_pcm_writei() to write fragments (a packet of frames
that is no longer than 1/4 the ring buffer) without checking the state
of the ring buffer first. Then if we get EAGAIN we go to sleep for as
long as it would take to play the fragment. This approach makes two
assumptions:

1) we will only get EAGAIN when there's no room on the ring
buffer to write the current fragment (so sleeping for the length of time
it would take to play the current packet should not cause underrun since
the packet is 1/4 the size of the ring buffer or less.

2) That if we get EAGAIN no frames where written at all.

I'm not sure which of these is not true (the documentation is not clear)
but with that approach the ring buffer gets corrupted (sounds like
samples are being played out of order). The solution that worked is to
check the amount of space available on the ring buffer before writting
frames. If there's no enough room then we sleep for just long enough for
there to be enough room in the buffer to write the whole fragment. This
method is better at preventing underruns (since we write the fragment as
soon as there is room) and also prevents ring buffer corruption.

This commit also tweaks QEMU audio settings, however there are still
audio problems when using the emulator which I assume are related to the
emulator settings since it works flawlessly on real hardware.