audio: Handle ALSA errors correctly
Keeping track of the stream time through ALSA timestamps and
snd_pcm_avail() turns out to be harder than it seems (impossible to do
it reliably AFAICT). So we're counting frames now and calculate the
runtime based on the number of frames when the street stops. While the
stream is playing we still rely on ALSA timestamps