player: Fix race condition between audio and video threads
Since the clock source is initialized when the audio thread is started
we need to start the audio thread first. Otherwise the video output
thread may start before the clock source is set to the audio clock and
it may go into a frame dropping spree or deadlock when the clock source
is changed.