audio: Fix audio glitches
This commit attempts to fix the audio glitches. It seems to have fixed
the glitchy sound issues on real hardware but under QEMU the glitches
still happen, though less often. That and the fact that the next commit
switches to using blocking mode the way we where using it before the
glitches started (on real hardware) leads me to believe that the
remaining issues are QEMU related. So we're keeping this commit as a
reference as we may wish to go back to it once we fully understand the
problem.

Note: What seems to have made a difference in this commit was the
addition to the snd_pcm_avail() call on avbox_audiostream_gettime() and
perhaps also prior to calling snd_pcm_wait(). I'm not sure if proper
handling of snd_pcm_wait() return codes and the lower timeout value also
made a difference. More testing and research is needed.