process: Only create STDOUT and STDERR pipes when needed
We were creating pipes for all child processes' stdout and stderr file
descriptor. However after a recent commit we're only reading from the
pipe for processes that are started with the AVBOX_PROCESS_STDIO_xxx and
AVBOX_PROCESS_STDERR_xxx flags. This was done to fix a high cpu usage
problem (aparently some process was writing a lot to stdout and we were
wasting cycles reading it). So now once a process fills the stdout pipe
it blocks forever.

This commit fixes the problem by creating the stdio pipes only for
processes that needed. For all other processes we set stderr and stdout
to /dev/null.