audio: Use ALSA non-blocking mode
Use non-blocking writes. This adds a little more latency to
avbox_audiostream_pause() but it solves the glitchy audio issues (at
least on real hardware) and it's simpler. We also limit the write size
to 256 frames to minimize latency.

This commit also switches to using avbox_queue instead of a linked list
with it's own synchronization code.

NOTE: Audio glitches still happen on QEMU but I haven't been able to
reproduce them in real hardware.