mediabox: Add message dispatcher
The main purpose of this commit was to replace the old messaging and
input systems based on pipes with a new fully-asynchronous message
dispatcher but doing this required rewriting large parts of the system.
The source code was also re-organized in order to make it easier to
split into a separate library and executable in the future. It would've
been messy to split it into a separate commit since the system was
totally broken at the time it was done.

In summary this commit:

- Adds an in-memory message dispatch
- Rewrites all UI components to use the new async model
- Adds a thread pool
- Moves all user interface code to the main thread
- Removes unbound latencies from main thread
- Fixed limitations on compositor and adds the option to
  create a dispatch object to windows (tying input to windows)