sdk: Use local tree as build source
Use the local tree (should be a git repo) as the source for the mediabox
package. So it doesn't need to be uploaded to bitbucket in order to
rebuild, thus enabling offline development.

Also disable the caching of the source so that the (local) git repo is
cloned everytime we rebuild (otherwise it is cached as a tarball that
needs to be deleted before new changes take effect).