player: Get rid of useless copying of video frames
Instead of queueing a copy of the decoded frame for the video output
thread we know queue a reference to the AVFrame, saving the unecessary
frame copy. Also we did away with the preallocated frame buffers and we
now allocate a new AVFrame for each frame. We also moved the scaling
from the filtergraph to the output thread.

Besides saving the frame copy this commit also reduces tearing as (at
least some of it) was being cause by overruning the frames buffer. The
next step is to add an avbox_window_blitscale() function to the video
driver and use that for scaling instead, that way if we use a hardware
accelerated driver (DirectFB with the right drivers) the scaling will be
accelerated (which paves the road for an accelerated decoder).