dispatch: Fix destructor semantics
After switching to message based destructors we we're not actually
marking the object as destroyed until the cleanup phase. The result of
this is that objects can receive messages after receiving the DESTROY
message. This commit fixes the issue by marking the object as destroyed
during the handling of the DESTROY message.

TODO: Right now we're printing a warning everytime a message is received
for a destroyed object. Maybe we should add a NOLEAK flag to messages in
order to supress the warning.