player: Implement control thread and async API
This commit is almost a rewrite of the media player widget. It now
spawns a dedicated thread to manage the playback threads and respond to
errors on the worker threads. All player commands (play, pause, stop,
seek, etc) are now asynchronous with a void return type. The application
must listen for player status messages to determine if the call
succeded.

This new structure makes the player code much easier to understand and
makes it significantly more stable and predictable. A number of bugs
that where too difficult to debug/reproduce with the old structure have
already been fixed in this commit.

The player is now working better than before but there is still a
possible memory leak when a decoding thread bails out unexpectedly any
packets that are enqueued after the thread bails out get leaked.