dispath: Implement delayed destructor
When an object receives the DESTROY message it can now return
AVBOX_DISPATCH_CONTINUE to delay the destruction of the object. In such
case the DESTROY message will be sent again after a delay of 500ms.

This is useful when you need to wait on another thread in the
destructor.