library: Fix race condition
The worker that populates the list was detached so if the library is
closed while it still running it would result in a SEGFAULT, and if a
subdirectory is selected while the worker is till running it would
result in items from the first directory added to the new directory's
list. This commit fixes it by waiting for the worker to exit before
selecting a new directory or destroying the library.

This works but it's way too much code and way too ugly. So we need the
ability to delay ANY message (not just DESTROY) on the dispatcher.