android: Added the option to send a URL to from the remote app
This commit adds a "Send URL" option on the menu that when tapped
displays a dialog with a text field. If you enter a URL and tap "Send"
the URL will be send to MediaBox.

Currently mediabox will see this as an invalid command. But the plan is
to implement it so that the URL is downloaded or streamed.