player: Fix race condition between pause and buffering messages
Prior to this commit if we try to pause while buffering we'll deadlock
because the checkpoint was not designed to be re-entrant (ie. two
consecutive calls to avbox_checkpoint_halt() would cause the 2nd call to
block forever. This commit fixes the problem by rewriting the checkpoint
to be re-entrant (ie. two consecutive calls to avbox_checkpoint_halt()
now require two calls to avbox_checkpoint_continue()).