library: Separate library backend and UI code
The library backend (renamed to library.c) now exports an API similar to
POSIX readdir() API to query a unified media library. Now we can have
different types of sources (local, UPnP, Bluetooth, etc) listed in the
media library. It also paves the way for a plugin system.

The library UI (now browser.c) now uses the library backend to populate
it's list and has been renamed to browser.c since it's now just a UI to
browse the library.