library: Implement local files library
This commit implements the local library. It adds the command line
parameter --store=/dev/sdX used to specify the media partition. This
partition is mounted by mediabox so it must not be mounted before
starting mediabox.

This is a WIP. It works but the code is messy since it combines code
that belongs on different files. Also during the initial scan all reads
fail due to the database being constantly locked.