player: Fix scaling issues and playback issues
We cannot rely on the fact that if the input video width is larger than
the height then the output with will also be larger. After applying the
aspect ratio adjustments it may not be so. We fix this problem by always
checking after the adjustment and if it doesn't turn out to be as
expected then recalculate fixing the other dimension. There's probably a
better way to do this but since it's only done once when playback is
started it is no big deal.

This commit also offsets the video to the right when the width is
smaller than the screen width.

It also fixes a buffer overrun if the video player returns the wrong
duration and increases the buffer size (again) as some videos still
deadlock with a smaller buffer.