player: Cleanup and audio stream change support
This commit fixes issues with DVD audio track switching and it also adds
support for track switching on regular stream (non-DVD). Besides that a
lot of cleanup was done to the player code. This works no only through
the DVD menus but through the remote's TRACK button (used to be
SUBTITLES but I just renamed as in the future the long press will be
used to switch subtitle tracks).