player: Fix short freezups after pause and cleanup still frame code
Before this commit the decoders where designed to automatically flush
themselves if the input stream underruns. The idea being that if the
input stream underruns and any frames needed to completely flush the
pipelines (since the audio drives the video we may need all audio frames
out before we can flush the video). The problem with that is that when
the decoder is flushed all frames until the next I-frame are lost and
this results in a freezeup of the video for up to several seconds.

This commit modifies the decoder to never flush unless the *flushing*
flag indicates that the specific decoder should be flushed. This flag is
only set by avbox_player_flush() and avbox_player_drop().

This commit also ensures that the still_frame condition is properly
cleared if we stop during a still frame.