player: Optimize audio track switch
This commit optimizes the audio track switch for speed. It modified the
audio decoder to be self synchronizing so the SET_CLOCK messages are now
meaningless for streams with audio. The next step is to make the video
decoder also self-synchronizing and get rid of those messages.

Also the audiostream now doesn't need to be flushed in order to update
the clock. It is now done through a control packet that most travel the
pipeline before taking effect.

Taken together it means that changing audio tracks is as simple as
setting the audio_stream_index variable and letting the decoder take
care of it.