player: Fix huge latency in menus with still video
Some DVD menus have a single video frame and a lot of audio so the one
video frame may stay stucked in the decoder buffers since now we only
flush on command. The solution is to have the video output thread send
an underrun message to the control twice a second while the decoded
frames buffer is empty. The control thread responds by telling the
decoder thread to flush it's decoder buffers.