video: Add accelerated X11 driver
This commit adds an accelerated (opengl) video driver for the X server.
Most of the code for this driver (the GL part) will also be used by the
DRM driver. The only reason to support the X server is because some
things like modesetting, scaling, overscan, etc. are only fully
supported by the X server. For systems where the display works out of
the box with drm (you can get a framebuffer console at full resolution)
the drm driver will be a better choice (once it supports GL).