video-drm: Add OpenGL support to drm driver
This commit splits the X11 driver into an X11 mini-driver and a pure
OpenGL driver on top of it. It also re-uses the OpenGL driver to add
acceleration to the libdrm driver.

Currently the DRM driver initializes the GL driver after it has done
it's own initialization. Which includes creating a root surface that
will never be used once the GL driver starts. In the future we should do
the GL initialization first or at least release the root surface once
the GL driver starts. We did it this way because it was easier since the
creation of the root surface is currently done during modesetting.

When no hardware acceleration is available the DRM driver will still use
OpenGL with the software renderer (llvmpipe) which is way slower than
our own software renderer so the --no-accel command-line option was
added for those cases. I will consider automagically falling back to our
software renderer in the future. For now this is more convenient for
debugging purposes.