player: Fix a few bugs that surfaced with GL driver
This commit fixes a few bugs that surfaced when we switched to the
OpenGL driver. The most significant is that when a multicast or anycast
message is sent it is dispatched to the thread of the first recipient.
We worked around it by opening the DVD stream from the main thread. This
is only a workaround and will cause the main thread to block while
opening a DVD. The real problem with the message dispatcher needs to be
fixed ASAP.