video-drm: Split drm driver into modesetting and software
This commit splits the software rendering part of the DRM driver into a
file of it's own that can be re-used with other drivers. This will
allows us to use it as a fallback on the X11 driver if GL acceleration
is not available. It'll probably take a while as I'm not too excited
about a native X11 driver when I have an OpenGL one :)