video-opengl: Offload YUV to RGB conversion to GPU
Instead of using software filters to convert YUV videos to RGB we now do
the conversion on the GPU with the OpenGL driver. The software driver
uses libswscale to do the conversion on software. The directfb driver
should also work (and be accelerated if you have the right drivers) but
I have not had a chance to test it yet.