player: Add bittorrent streaming support
Add BitTorrent streaming support. Now when you send a URL from the
android app you are given to choice to download or stream. If stream is
chosen the largest file on the torrent will be downloaded and streamed.
You can also send http URLs to torrent files but that feature has not
been well tested.

Still TODO:

1. Check the integrity of streamed content. When streaming torrents all
   files play correctly but ffmpeg reports errors and we don't get the
   first video frame until after about 12 seconds in.

2. Add the ability to stream specific files to the torrent, not just
   the latest. We also want the ability to get a substream of the same
   torrent. That will be used by the subtitles API to find subtitles in
   the torrent.