player: Fine tune for RT
There is too much on this commit to list here. It is the result of lots
of profiling to fine tuning to perform well with RT scheduling. Maybe
the biggest change is that all memory used during playback is now
pooled. Also includes performance improvements to the mmal decoder and
the whole player stack.