ALSA: Always handle ALSA errors immediately
Previously we were only handling ALSA errors (attempting to recover from
underrun) when the call to snd_pcm_writei() and ignoring them everywhere
else since eventually it would be called. There is nothing on ALSA
documentation saying this is wrong, other functions should continue to
fail until recovery is done. However, the ALSA examples do it
differently and I believe that may be the reason why snd_pcm_wait() was
sporadically deadlocking on prior revisions that used it. We solved that
problem poorly by sleeping the amount of time that it would take before
there is room in the ring buffer. This works on most system but it can
lead to underruns if we don't wake up in time.

This commit attempts to solve all these issues by switching back to
using snd_pcm_wait() and handling underruns immediately after they are
detected. It will need to be tested thoroughly before it can be
considered to have solved the problem since it happened so randomly.